Fix critical code review issues in straggler mitigation
Address 8 critical and important implementation issues:

P0 Critical Issues:
1. Sequential semaphore lifecycle: Remove readonly and disposal to support
   restart scenarios without ObjectDisposedException
2. Sequential semaphore TOCTOU race: Capture mode atomically at acquisition
   time to prevent semaphore count drift
3. Try/finally coverage: Move metrics initialization inside try block to
   ensure cleanup always runs and prevent memory leaks

P1 Important Issues:
4. Duplicate straggler detection: Add tracking dictionary to prevent counting
   same file multiple times across retry cycles
5. Counter overflow protection: Change from int to long (max ~9 quintillion)
   to prevent overflow in pathological scenarios

P2 Issues:
6. Cleanup task cancellation: Use cancellationToken in fire-and-forget cleanup
   to respect shutdown and remove immediately if cancelled
7. File size validation: Add constructor validation to reject zero or negative
   file sizes and prevent invalid throughput calculations
8. Stale CTS atomicity: Use AddOrUpdate to atomically replace cancellation
   token source and dispose old one, preventing race conditions

All 18 straggler mitigation tests pass after fixes.

+++ csharp/test/Drivers/Databricks/E2E/CloudFetch/CloudFetchStragglerE2ETests.cs
@@ -126,7 +126,7 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Databricks.E2E.CloudFetch
             var stragglers = detector.IdentifyStragglerDownloads(metrics, System.DateTime.UtcNow.AddSeconds(10));
 
             // Assert - Counter should increment for detected stragglers
-            int count = detector.GetTotalStragglersDetectedInQuery();
+            long count = detector.GetTotalStragglersDetectedInQuery();
             Assert.True(count >= 0); // Counter should be non-negative
         }
     }
